Extract pretty formating of time into separate callable function.
Since needed at two places. The other place being a RAISE NOTICE
also added in this commit to show the number of executinos so far
and the intermediate result of the measured execution time,
useful to get a feeling for when it's pointless to continue waiting
for N number of sig digits, if observing the value doesn't converge.