Add min_time with a 10 ms default. Also add a new reporting view.
Experimentation have shown that a minimum execution time of 10 ms
works good to prevent getting a final result with the requested
numbers of significant digits by accident.

Also move the timeout code to the end of the loop, to consider if
we should continue only after first checking if we can produce
a result. Otherwise we could timeout even though already having a
result.