Oonly look at last three values in regression computation.
Since we're not interested in the potentially noisy measurements
when the number of executions were too low, we only want to look
at the last ones, and conclude if they lay on a more or less
straight line. Also, a default R^2 threshold of 0.99 seems
sufficient, 0.999 might be an impossible task on some CPUs.